requirements.txt
Pillow==12.0.0
pystray==0.19.5
Requests==2.32.5
